+++ common/db/src/main/java/is/codion/common/db/operation/FunctionType.java
@@ -20,7 +20,7 @@ package is.codion.common.db.operation;
 
 /**
  * @param <C> the connection type
- * @param <T> the function argument type
+ * @param <T> the function parameter type
  * @param <R> the function result type
  */
 public interface FunctionType<C, T, R> {
@@ -34,7 +34,7 @@ public interface FunctionType<C, T, R> {
 	 * Creates a {@link FunctionType} with the given name and types.
 	 * @param name the name
 	 * @param <C> the connection type
-	 * @param <T> the function argument type
+	 * @param <T> the function parameter type
 	 * @param <R> the function result type
 	 * @return a new {@link FunctionType}
 	 */